Fix leaking MapView due to annotations plugin (#115)
* Fix leaking MapView due to annotations plugin

When using SymbolManager (part of the annotations plugin), a strong
reference is kept to the MapView until it is explicitly cleared. This
patch adds methods to allow calling AnnotationManager's onDestroy to
explicitly clear the MapView reference.

* Annotate callers with @UiThread

AnnotationManager's onDestroy method is annotated as @UiThread, so
explicitly annotate callers of it as @UiThread.